MultipleLocationShipPlacer: works out all the possible locations for a ship
MultipleLocationShipPlacer is needed for working out the possible scenarios in a probabilistic shooter. It is a modified clone of ShipLocationChooser.
Functionally chooser.choose(findAllValidLocations(...)) is equivalent to new ShipLocationChooser(chooser).place(2, grid), but the former is much slower. I think they could be brought back together without paying a humongous performance by using a monad or an Iterator, but I need to investigate.